+++ pingouin/power.py
@@ -4,8 +4,8 @@ import numpy as np
 from scipy import stats
 from scipy.optimize import brenth
 
-__all__ = ["power_ttest", "power_ttest2n", "power_anova", "power_corr",
-           "power_chi2"]
+__all__ = ["power_ttest", "power_ttest2n", "power_anova", "power_rm_anova",
+           "power_corr", "power_chi2"]
 
 
 def power_ttest(d=None, n=None, power=None, alpha=0.05, contrast='two-samples',
@@ -346,7 +346,7 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
     Parameters
     ----------
     eta : float
-        ANOVA effect size (eta-square == :math:`\\eta^2`).
+        ANOVA effect size (eta-square = :math:`\\eta^2`).
     k : int
         Number of groups
     n : int
@@ -355,7 +355,7 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
     power : float
         Test power (= 1 - type II error).
     alpha : float
-        Significance level (type I error probability).
+        Significance level :math:`\\alpha` (type I error probability).
         The default is 0.05.
 
     Notes
@@ -378,12 +378,20 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
     Statistical power is mainly affected by the effect size and the sample
     size.
 
-    For one-way ANOVA, eta-square is the same as partial eta-square. It can be
-    evaluated from the f-value and degrees of freedom of the ANOVA using
-    the following formula:
+    For one-way ANOVA, eta-square is the same as partial
+    eta-square. It can be evaluated from the F-value (:math:`F^*`) and the
+    degrees of freedom of the ANOVA (:math:`v_1, v_2`) using the following
+    formula:
 
-    .. math::
-        \\eta^2 = \\frac{v_1 F^*}{v_1 F^* + v_2}
+    .. math:: \\eta^2 = \\frac{v_1 F^*}{v_1 F^* + v_2}
+
+    Note that GPower uses the :math:`f` effect size instead of the
+    :math:`\\eta^2`. The formula to convert from one to the other are given
+    below:
+
+    .. math:: f = \\sqrt{\\frac{\\eta^2}{1 - \\eta^2}}
+
+    .. math:: \\eta^2 = \\frac{f^2}{1 + f^2}
 
     Using :math:`\\eta^2` and the total sample size :math:`N`, the
     non-centrality parameter is defined by:
@@ -393,7 +401,7 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
     Then the critical value of the non-central F-distribution is computed using
     the percentile point function of the F-distribution with:
 
-    .. math:: q = 1 - alpha
+    .. math:: q = 1 - \\alpha
     .. math:: v_1 = k - 1
     .. math:: v_2 = N - k
 
@@ -407,11 +415,10 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
     variables (i.e. sample size, effect size, or significance level). If the
     solving fails, a nan value is returned.
 
-    Results have been tested against GPower and the R pwr package.
+    Results have been validated against GPower and the R pwr package.
 
     References
     ----------
-
     .. [1] Cohen, J. (1988). Statistical power analysis for the behavioral
            sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
 
@@ -476,7 +483,7 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
     elif k is None:
         # Compute required number of groups
 
-        def _eval_k(k, eta, n, power, alpha):
+        def _eval_k(k, f_sq, n, power, alpha):
             return func(f_sq, k, n, power, alpha) - power
 
         try:
@@ -517,6 +524,206 @@ def power_anova(eta=None, k=None, n=None, power=None, alpha=0.05):
         try:
             return brenth(_eval_alpha, 1e-10, 1 - 1e-10, args=(f_sq, k, n,
                                                                power))
+        except ValueError:  # pragma: no cover
+            return np.nan
+
+
+def power_rm_anova(eta=None, m=None, n=None, power=None, alpha=0.05,
+                   corr=0.5, epsilon=1):
+    """
+    Evaluate power, sample size, effect size or
+    significance level of a balanced one-way repeated measures ANOVA.
+
+    Parameters
+    ----------
+    eta : float
+        ANOVA effect size (eta-square = :math:`\\eta^2`).
+    m : int
+        Number of repeated measurements.
+    n : int
+        Sample size per measurement. All measurements must have the same
+        sample size.
+    power : float
+        Test power (= 1 - type II error).
+    alpha : float
+        Significance level :math:`\\alpha` (type I error probability).
+        The default is 0.05.
+    corr : float
+        Average correlation coefficient among repeated measurements.
+        The default is :math:`r=0.5`.
+    epsilon : float
+        Epsilon adjustement factor for sphericity. This can be
+        calculated using the :py:func:`pingouin.epsilon` function.
+
+    Notes
+    -----
+    Exactly ONE of the parameters ``eta``, ``m``, ``n``, ``power`` and
+    ``alpha`` must be passed as None, and that parameter is determined from
+    the others.
+
+    Notice that ``alpha`` has a default value of 0.05 so None must be
+    explicitly passed if you want to compute it.
+
+    Statistical power is the likelihood that a study will
+    detect an effect when there is an effect there to be detected.
+    A high statistical power means that there is a low probability of
+    concluding that there is no effect when there is one.
+    Statistical power is mainly affected by the effect size and the sample
+    size.
+
+    For one-way repeated measure ANOVA, eta-square is the same as partial
+    eta-square. It can be evaluated from the F-value (:math:`F^*`) and the
+    degrees of freedom of the ANOVA (:math:`v_1, v_2`) using the following
+    formula:
+
+    .. math:: \\eta^2 = \\frac{v_1 F^*}{v_1 F^* + v_2}
+
+    Note that GPower uses the :math:`f` effect size instead of the
+    :math:`\\eta^2`. The formula to convert from one to the other are given
+    below:
+
+    .. math:: f = \\sqrt{\\frac{\\eta^2}{1 - \\eta^2}}
+
+    .. math:: \\eta^2 = \\frac{f^2}{1 + f^2}
+
+    Using :math:`\\eta^2`, the sample size :math:`N`, the number of repeated
+    measurements :math:`m`, the epsilon correction factor :math:`\\epsilon`
+    (see :py:func:`pingouin.epsilon`), and the average correlation between
+    the repeated measures :math:`c`, one can then calculate the
+    non-centrality parameter as follow:
+
+    .. math:: \\delta = \\frac{f^2 * N * m * \\epsilon}{1 - c}
+
+    Then the critical value of the non-central F-distribution is computed using
+    the percentile point function of the F-distribution with:
+
+    .. math:: q = 1 - \\alpha
+    .. math:: v_1 = (m - 1) * \\epsilon
+    .. math:: v_2 = (N - 1) * v_1
+
+    Finally, the power of the ANOVA is calculated using the survival function
+    of the non-central F-distribution using the previously computed critical
+    value, non-centrality parameter, and degrees of freedom.
+
+    :py:func:`scipy.optimize.brenth` is used to solve power equations for other
+    variables (i.e. sample size, effect size, or significance level). If the
+    solving fails, a nan value is returned.
+
+    Results have been validated against GPower.
+
+    References
+    ----------
+    .. [1] Cohen, J. (1988). Statistical power analysis for the behavioral
+           sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
+
+    .. [2] https://cran.r-project.org/web/packages/pwr/pwr.pdf
+
+    Examples
+    --------
+    1. Compute achieved power
+
+    >>> from pingouin import power_rm_anova
+    >>> print('power: %.4f' % power_rm_anova(eta=0.1, m=3, n=20))
+    power: 0.8913
+
+    2. Compute required number of groups
+
+    >>> print('m: %.4f' % power_rm_anova(eta=0.1, n=20, power=0.90))
+    m: 3.1347
+
+    3. Compute required sample size
+
+    >>> print('n: %.4f' % power_rm_anova(eta=0.1, m=3, power=0.80))
+    n: 15.9979
+
+    4. Compute achieved effect size
+
+    >>> print('eta: %.4f' % power_rm_anova(n=20, m=4, power=0.80, alpha=0.05))
+    eta: 0.0680
+
+    5. Compute achieved alpha (significance)
+
+    >>> print('alpha: %.4f' % power_rm_anova(eta=0.1, n=20, m=4, power=0.80,
+    ...                                   alpha=None))
+    alpha: 0.0081
+    """
+    # Check the number of arguments that are None
+    n_none = sum([v is None for v in [eta, m, n, power, alpha]])
+    if n_none != 1:
+        msg = 'Exactly one of eta, m, n, power, and alpha must be None.'
+        raise ValueError(msg)
+
+    # Safety checks
+    assert 0 < epsilon <= 1, 'epsilon must be between 0 and 1.'
+    assert 0 < corr < 1, 'corr must be between 0 and 1.'
+    if eta is not None:
+        eta = abs(eta)
+        f_sq = eta / (1 - eta)
+    if alpha is not None:
+        assert 0 < alpha <= 1, 'alpha must be between 0 and 1.'
+    if power is not None:
+        assert 0 < power <= 1, 'power must be between 0 and 1.'
+    if n is not None:
+        assert n > 1, 'The sample size n must be > 1.'
+    if m is not None:
+        assert m > 1, 'The number of repeated measures m must be > 1.'
+
+    def func(f_sq, m, n, power, alpha, corr):
+        dof1 = (m - 1) * epsilon
+        dof2 = (n - 1) * dof1
+        nc = (f_sq * n * m * epsilon) / (1 - corr)
+        fcrit = stats.f.ppf(1 - alpha, dof1, dof2)
+        return stats.ncf.sf(fcrit, dof1, dof2, nc)
+
+    # Evaluate missing variable
+    if power is None:
+        # Compute achieved power
+        return func(f_sq, m, n, power, alpha, corr)
+
+    elif m is None:
+        # Compute required number of repeated measures
+
+        def _eval_m(m, f_sq, n, power, alpha, corr):
+            return func(f_sq, m, n, power, alpha, corr) - power
+
+        try:
+            return brenth(_eval_m, 2, 100, args=(f_sq, n, power, alpha, corr))
+        except ValueError:
+            return np.nan
+
+    elif n is None:
+        # Compute required sample size
+
+        def _eval_n(n, f_sq, m, power, alpha, corr):
+            return func(f_sq, m, n, power, alpha, corr) - power
+
+        try:
+            return brenth(_eval_n, 5, 1e+6, args=(f_sq, m, power, alpha, corr))
+        except ValueError:
+            return np.nan
+
+    elif eta is None:
+        # Compute achieved eta
+
+        def _eval_eta(f_sq, m, n, power, alpha, corr):
+            return func(f_sq, m, n, power, alpha, corr) - power
+
+        try:
+            f_sq = brenth(_eval_eta, 1e-10, 1 - 1e-10, args=(m, n, power,
+                                                             alpha, corr))
+            return f_sq / (f_sq + 1)  # Return eta-square
+        except ValueError:  # pragma: no cover
+            return np.nan
+
+    else:
+        # Compute achieved alpha
+
+        def _eval_alpha(alpha, f_sq, m, n, power, corr):
+            return func(f_sq, m, n, power, alpha, corr) - power
+
+        try:
+            return brenth(_eval_alpha, 1e-10, 1 - 1e-10, args=(f_sq, m, n,
+                                                               power, corr))
         except ValueError:  # pragma: no cover
             return np.nan