+++ docs/changelog.rst
@@ -14,6 +14,9 @@ v0.2.7 (July 2019)
 a. Added support for long-format dataframe in :py:func:`pingouin.sphericity` and :py:func:`pingouin.epsilon`.
 b. Added support for two within-factors interaction in :py:func:`pingouin.sphericity` and :py:func:`pingouin.epsilon` (for the former, granted that at least one of them has no more than two levels.)
 
+**New functions**
+
+a. Added :py:func:`pingouin.power_rm_anova` function.
 
 v0.2.6 (June 2019)
 ------------------